:recycle: Factor out extraction into standalone Extractor
